JL8-20 refactor : 플레이 리스트 아이템 위치 이동
